Require recent App-cpanminus-reporter.
This module subclasses App::cpanminus::reporter and, as such, is sensitive to
the version of that library which a user has installed.  I have received
CPANtesters reports which FAIL this module, but so far in each case an older
(0.12 or 0.13) version of App::cpanminus::reporter was being used.  The
Changes file for App-cpanminus-reporter indicates that the handling of
UNKNOWN/NA reports changed in version 0.14 -- and the unit tests that were
failing fell into this category.  So I'm going to require that users upgrade
to the same version of App-cpanminus-reporter that I used during development:
0.17.