Start work on transmit_report().
Objective:  Both retain a report on disk and transmit a report
to CPANtesters (as we would if we were just using cpanm-reporter).

Assuming that we are still inheriting our constructor from
App::cpanminus::reporter, then we need to inject a KVP into the
CPAN::cpanminus::reporter::RetainReports objects after it has been constructed
but before run() is called.  We then need to adapt make_report() to use that
KVP to transmit a report.

TODO: In doing so, we need values for certain KVPs to
pass to a Test::Reporter call.  For those we'll have to re-study
App::cpanminus::reporter.