No longer need my own get_author() method
App::cpanminus::reporter modified get_author() in its 0.19 release.
This was done to accommodate processing of the 'file' scheme and was
accompanied by additional tests.  This obviated the need for
CPAN::cpanminus::reporter::RetainReports to define its own get_author()
method, but required modification of tests to accommodate the 'file'
scheme.

It should be noted that App::cpanminus::reporter is intended to generate
a report to an author, so 'author' must have a defined value.  This
value gets defined by parsing a URI like
'/authors/id/M/ME/METATEST/Phony-PASS-0.01.tar.gz', where the 'author'
is the author's CPAN ID, in this case 'METATEST'.  This meant that I had
to move a tarball used for testing from 't/data/' to
't/data/M/ME/METATEST/'.