Branch to expose problem ONLY; DO NOT MERGE
Add debugging code to RetainReports.pm.  Strip down 001-load.t to just
the block testing the 'file' URI scheme.  (Don't bother running any
tests other than 001.)

The problem is that App::cpanminus::reporter::parse_uri absolutely
expects to find a tarball at the bottom of a directory tree that looks
like https://www.cpan.org, i.e., a 3-tier structure like:

        .../M/ME/METATEST/Phony-PASS-0.001.tar.gz

So if you are building a library which inherits from
App::cpanminus::reporter (as this library does), and if you want to
include a dummy tarball for testing, you will have to situate that
tarball someplace like this:

        .../top-level-dir/t/data/M/ME/METATEST/Phony-PASS-0.001.tar.gz

... rather than like this:

        .../top-level-dir/t/data/Phony-PASS-0.001.tar.gz

... in order to get parse_uri() to correctly extract 'author'.  Failure
to make this extraction will cause $reporter->run to return without a
defined value and no JSON files will be created for this or any other
tarball.